requirement.txt
pandas-datareader
scipy-stack
numpy
matplotlib
pytrends
goose3
bs4
textblob
nltk
backtrader
